requirements.txt
Django==4.0.3
djongo==1.3.6
pymongo==3.12.3
djangorestframework==3.13.1
requests==2.27.1
django-cors-headers==3.11.0
# Removed qrcode and Pillow which were causing installation issues
